changed file name and enhanced code for connection with the processor repo
